tk-15 : implentation of dvd preview based on database

+++ Models/Database/Loan.cs
@@ -27,6 +27,6 @@ namespace RopeyDVDManagementSystem.Models
 
         public DateTime DateReturned { get; set; }
 
-        public uint ReturnAmount { get; set; }
+        public decimal ReturnAmount { get; set; }
     }
 }
